refactor: seed 로직을 lib/airflow/bootstrap 패키지로 분리
db/ 패키지가 DB 연결·마이그레이션 외에 Swagger 파싱, HTTP fetch,
예시 데이터 생성, Init 오케스트레이션까지 담당하던 God Package 문제를
해소.

- lib/airflow/bootstrap/swagger.go: Swagger 2.0 파싱 타입과 변환 로직
- lib/airflow/bootstrap/taskcomponent.go: TaskComponentInit 오케스트레이터
  (외부 모듈 Swagger → TaskComponent 계약 동기화 파이프라인)
- lib/airflow/bootstrap/workflowtemplate.go: WorkflowTemplateInit 로더
- cmd/cm-cicada/main.go: Init 호출을 db.Open() 에서 main으로 이관.
  LoadExamples 플래그 검사도 main에서 수행
- db/taskComponent.go, db/workflowTemplate.go 삭제
- db/sqlite.go: Init 호출 제거, DB 열기/마이그레이션만 담당하도록 정리
- controller: db.XxxGetByName 호출을 dao.XxxGetByName 으로 전환

db/compat.go 에 TaskComponentGetByName shim을 남겨 lib/airflow/gusty.go
의 잔여 참조가 깨지지 않도록 유지. 다음 커밋에서 gusty.go 를 dao로
마이그레이션한 뒤 shim 을 제거할 예정.
